SDN-1569: Add support for multiple destinations in redirect mode
Presently if more than one destination is specified through the
"destinations" array in the NAD, we only pick the first destination
and ignore the rest. This patch adds support for multiple
destinations.

Also in OCP 3.11 we used to support multiple destination specification
in 3 formats:

1)["ip-address/mask"]
2)["port protocol ip-address/mask"]
3)["port protocol ip-address/mask remote-port"]

in multiple lines. We do the same for feature parity.